refactor: small fix for sql schema
