+++ docker-compose.yaml
@@ -19,11 +19,11 @@ services:
       - db
     environment:
       - SPRING_DATASOURCE_URL=jdbc:postgresql://db:5432/shareit
-      - POSTGRES_USER=root
-      - POSTGRES_PASSWORD=root
+      - POSTGRES_USER=shareit
+      - POSTGRES_PASSWORD=shareitpswd
 
   db:
-    image: postgres:12.14-alpine
+    image: postgres:14-alpine
     container_name: shareit_db_container
     volumes:
       - /var/lib/postgresql/data/
@@ -31,5 +31,5 @@ services:
       - "6541:5432"
     environment:
       - POSTGRES_DB=shareit
-      - POSTGRES_USER=root
-      - POSTGRES_PASSWORD=root
+      - POSTGRES_USER=shareit
+      - POSTGRES_PASSWORD=shareitpswd
