+++ gateway/pom.xml
@@ -51,13 +51,13 @@
             <optional>true</optional>
         </dependency>
 
+        <!--  Test  -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-test</artifactId>
             <scope>test</scope>
         </dependency>
 
-        <!--  Test  -->
         <dependency>
             <groupId>org.codehaus.groovy</groupId>
             <artifactId>groovy-all</artifactId>
@@ -82,5 +82,4 @@
             </plugin>
         </plugins>
     </build>
-
 </project>
