+++ server/src/main/resources/schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS requests (
                                      id INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
                                      description VARCHAR NOT NULL,
                                      requester_id INTEGER,
-                                     create_date TIMESTAMP,
+                                     create_date TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP NOT NULL,
                                      CONSTRAINT fk_requester_id FOREIGN KEY (requester_id) REFERENCES users (id) ON DELETE CASCADE
 );
 
@@ -25,8 +25,8 @@ CREATE TABLE IF NOT EXISTS items (
 
 CREATE TABLE IF NOT EXISTS bookings (
                                      id INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-                                     start_date TIMESTAMP NOT NULL,
-                                     end_date TIMESTAMP NOT NULL,
+                                     start_date TIMESTAMP WITH TIME ZONE NOT NULL,
+                                     end_date TIMESTAMP WITH TIME ZONE NOT NULL,
                                      status VARCHAR NOT NULL,
                                      item_id INTEGER,
                                      booker_id INTEGER,
@@ -42,7 +42,7 @@ CREATE TABLE IF NOT EXISTS comments (
                                      text VARCHAR NOT NULL,
                                      item_id INTEGER,
                                      author_id INTEGER,
-                                     created_date TIMESTAMP,
+                                     created_date TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP NOT NULL,
                                      CONSTRAINT fk_comments_item_id FOREIGN KEY (item_id) REFERENCES items (id) ON DELETE CASCADE,
                                      CONSTRAINT fk_author_id FOREIGN KEY (author_id) REFERENCES users (id) ON DELETE CASCADE
 );
